aws/lambdas/upload-avatar/requirements.txt
boto3==1.38.1
botocore==1.38.1
certifi==2025.1.31
charset-normalizer==3.4.1
ecdsa==0.19.1
idna==3.10
jmespath==1.0.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
python-dateutil==2.9.0.post0
python-jose==3.4.0
requests==2.32.3
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
urllib3==2.4.0
